bsd-user: Fix riscv's target_cpu_loop
Old versions of this code wrote to info's members, but this was split
out into individual variables. However, the condition for when to call
force_sig_fault still checks the struct member, which is uninitialised.
Change it to check the actual variable in use, and delete the now unused
info variable.

Signed-off-by: Jessica Clarke <jrtc27@jrtc27.com>