bsd-user: Correctly handle NULL msg_control for sendmsg
FreeBSD's sendmsg requires msg_controllen to be valid, i.e. at least one
message, if msg_control is non-NULL, with no special casing for 0. This
is needed for pkg < 1.21, which used socketpair + sendmsg for its worker
processes for creating a local repository. For recvmsg this might not
matter but we might as well be careful and faithfully mirror the
pointer's NULL-ness still.

Signed-off-by: Jessica Clarke <jrtc27@jrtc27.com>