bsd-user: Use correct values for VERIFY_READ/VERIFY_WRITE
Historically these were 0 and 1 in linux-user, and were translated to
PAGE_* appropriately, but 68f77666649b changed them to be just the
PAGE_* values. However bsd-user was in a mixed state where the values
weren't translated but were still defined as 0/1, meaning VERIFY_READ
was just verifying the page existed, and thus would return true for
PROT_NONE pages, and VERIFY_WRITE was just verifying the page was
readable.

This fixes bsd-user segfaulting when creating a coredump as it thought
the stack guard page was readable.

Signed-off-by: Jessica Clarke <jrtc27@jrtc27.com>