bsd-user: Make stack executable for sigtramp
Currently our sigtramp code lives at the top of the stack rather than in
its own page, so make the stack executable for now so signal delivery
works.

I'm guessing that this is only needed as of cdf713085131 ("accel/tcg:
Properly implement get_page_addr_code for user-only") and that's why
older versions work (the current emulators/qemu-user-static-devel port
is based on an upstream around 6 months older than that commit).

Signed-off-by: Jessica Clarke <jrtc27@jrtc27.com>