bsd-user: Rearrange thr_new register initialisation
The target_cpu_clone_regs call is given the wrong argument (stack base
not initial stack pointer), but turns out to be overwritten anyway by
the later target_thread_set_upcall. The call to target_cpu_set_tls can
also be done in the new thread itself, rather than in the parent, so all
the code to set the thread's registers is done in one place. This more
closely aligns with the real kernel implementation.

Signed-off-by: Jessica Clarke <jrtc27@jrtc27.com>