bsd-user: Don't mess with user-supplied PC for exec and upcalls
If the input is invalid, we should fault rather than mask the error.
Moreover, on RISC-V, PC can be only 2-byte aligned if you have the 'C'
extension, which is part of our baseline, so masking out bit 1 is even
more wrong. In practice the entry point is 4 byte aligned, though.

Signed-off-by: Jessica Clarke <jrtc27@jrtc27.com>