Add proper conversion of the clock types;
Properly convert AT_XXX flags for some of the [foo]at() calls;

Fake lchmod(2), which is not possible on linux, just make sure
it's accessible and is a symlink.

Don't overflow target buffer in getdirentries() if the storage
for the struct is more compact on host as it's the case on
linux.